rename registration event to client registration event
