+++ common/src/main/java/org/vivecraft/api/client/event/VivecraftClientRegistrationEvent.java
@@ -1,14 +1,18 @@
-package org.vivecraft.api.client;
+package org.vivecraft.api.client.event;
+
+import org.vivecraft.api.client.InteractModule;
+import org.vivecraft.api.client.Tracker;
+import org.vivecraft.api.client.VRClientAPI;
 
 import java.util.function.Consumer;
 
 /**
- * The event that is fired for registering trackers and interact modules with Vivecraft. One can interact with this
- * event by registering a handler using {@link VRClientAPI#addRegistrationHandler(Consumer)}.
+ * The event that is fired on the client for registering trackers and interact modules with Vivecraft. One can interact
+ * with this event by registering a handler using {@link VRClientAPI#addClientRegistrationHandler(Consumer)}.
  *
  * @since 1.3.0
  */
-public interface VivecraftRegistrationEvent {
+public interface VivecraftClientRegistrationEvent {
 
     /**
      * Registers the given trackers to the list of all trackers to be run for the local player. See the documentation
